Tests/LibWeb/Text/expected/css/box-shadow-resolves-length-functions.txt

0 calc(5px - 10px) 0 calc(2px + 3px) => #000000ff 0px -5px 0px 5px
